hal_hw_interface:  Add hal_hw_interface.hal_offset_mgr module
This component manages joint offsets.  It defines a new type of ROS
HAL pin object, actually a pair of pins for use in combination with
the HAL RT `offset` component:  a `<joint>_offset` output
pin and a `<joint>_fb-in` input pin for connecting to the `offset`
component pins of the same names.  The offset pin is a redis_store pin
that persists its value across machine restarts.

When a single `zero_joint` command IO pin goes high, the component
copies the feedback pin value to the offset pin for each joint, and
resets the command pin.  If the `enable` BIT input pin is high, the
component will refuse to do this and issue a warning.

At start-up, the HAL configuration scripts should pull the
`load_offset` command IO pin high after the component is connected.
The component will then copy values from redis to the offset pin for
each joint, and reset the command pin.