Major overhaul of controller classes for better reliability with stop planning and error reporting
* Effectively ignore probe transitions that occur at the very first
update of a new trajectory. For example, a rising edge at the first
update will be treated like the move started with the probe active
(error for mode 0, 2, 3, ok for 4, 5, 6)
* Factor out goal completion / abort methods into the base class so goal
/ path tolerance checks can use the same code path to signal an abort
(and stop motion).
* Fix realtime conflict when an "empty" trajectory is sent (effectively
a stop command). Previously, setHoldPosition was called directly from
the non-RT side. Because the hold trajectory is pre-allocated, if the
controller is already executing a stop trajectory, it may have
unpredicatable results for setHoldPosition to overwrite that trajectory
in place. The specific case this fixes is an exception in publishState
where the current / desired times were outside of the range of a
ros::Duration. Now, a stop is commanded from the non-RT side with an
atomic flag that is checked at trajectory update. setHoldPosition should
always be called from RT only.